include starter-tools dir; rename command to make mcp-server-docs
